refact: controllers -> organized by function
