+++ lib/controllers/controllers.dart
@@ -1,7 +1,7 @@
-export 'firebase_controller.dart';
-export 'messaging_controller.dart';
+export 'firebase/firebase_controller.dart';
+export 'firebase/messaging_controller.dart';
 export 'pdf_state_controller.dart';
-export 'protocol_version_controller.dart';
+export 'local_storage/protocol_version_controller.dart';
 export 'search_controller.dart';
-export 'storage_controller.dart';
+export 'local_storage/storage_controller.dart';
 export 'theme_controller.dart';
