+++ lib/main.dart
@@ -2,7 +2,7 @@ import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:get_storage/get_storage.dart';
-import 'package:wvems_protocols/controllers/messaging_controller.dart';
+import 'package:wvems_protocols/controllers/firebase/messaging_controller.dart';
 import 'package:wvems_protocols/controllers/controllers.dart';
 import 'package:wvems_protocols/ui/views/views.dart';
